pyproject.toml
[tool.poetry]
name = "quantrion"
version = "0.1.0"
description = ""
authors = ["Daniel Fonnegra <dfonnegrag@gmail.com>"]

[tool.poetry.scripts]
quantrion = "quantrion.main:run"

[tool.poetry.dependencies]
python = ">=3.8,<3.12"
numpy = "^1.23.2"
jupyter = "^1.0.0"
python-dotenv = "^0.20.0"
httpx = "^0.23.0"
matplotlib = "^3.5.3"
pandas = "^1.4.3"
optuna = "^2.10.1"
websockets = "^10.3"
seaborn = "^0.11.2"
plotly = "^5.10.0"
PyYAML = "^6.0"
scikit-learn = "^1.1.2"
pydantic = "^1.10.2"

[tool.poetry.dev-dependencies]
pytest = "^7.1.2"
pytest-cov = "^3.0.0"
pytest-asyncio = "^0.19.0"
pytest-httpx = "^0.21.0"
black = {extras = ["jupyter"], version = "^22.6.0"}
isort = "^5.10.1"
multiprocess = "^0.70.13"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.isort]
profile = 'black'
multi_line_output = 3